# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get all payments",
                "operationId": "826dc6b24dd704607aa018a59207dd96",
                "responses": {
                    "200": {
                        "description": "A list of payments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAllPaymentsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create payment",
                "operationId": "8bd089442f313f34b467e2625ff83655",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paymentMethod": {
                                        "description": "Method of the payment",
                                        "type": "string"
                                    },
                                    "vehicle_id": {
                                        "description": "ID of the vehicle associated with the payment",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreatePaymentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get payment by ID",
                "operationId": "5ec820dc5a59156a714c50a583e10016",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetPaymentByIdResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payment"
                ],
                "summary": "Delete payment by ID",
                "operationId": "05cc141cec4b7acba5a54f72f1e511cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeletePaymentByIdResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/{id}/finalize": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Finalize payment",
                "operationId": "2021c0ce919a5edf3f22b04097fa38d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment to finalize",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment finalized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FinalizePaymentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreatePaymentResource": {
                "title": "Create Payment Resource",
                "description": "Resource returned when a payment is created",
                "properties": {
                    "status": {
                        "description": "Indicates if the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "A message describing the result of the operation",
                        "type": "string"
                    },
                    "payment": {
                        "properties": {
                            "id": {
                                "description": "ID of the payment",
                                "type": "integer"
                            },
                            "value": {
                                "description": "Value of the payment",
                                "type": "number",
                                "format": "float"
                            },
                            "value_in_reais": {
                                "description": "Value of the payment in Brazilian reais",
                                "type": "number",
                                "format": "float"
                            },
                            "dateTime": {
                                "description": "Date and time when the payment was registered",
                                "type": "string",
                                "format": "date-time"
                            },
                            "paymentMethod": {
                                "description": "Method used for the payment",
                                "type": "string"
                            },
                            "paid": {
                                "description": "Indicates if the payment has been made",
                                "type": "boolean"
                            },
                            "vehicle_id": {
                                "description": "ID of the vehicle associated with the payment",
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "DeletePaymentByIdResource": {
                "title": "Delete Payment By Id Resource",
                "description": "Resource returned when a payment is deleted",
                "properties": {
                    "status": {
                        "description": "Indicates if the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "A message describing the result of the operation",
                        "type": "string"
                    },
                    "payment": {
                        "description": "Details of the payment, which is null in this case",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "FinalizePaymentResource": {
                "title": "Finalize Payment Resource",
                "description": "Resource returned when a payment is finalized",
                "properties": {
                    "status": {
                        "description": "Indicates if the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "A message describing the result of the operation",
                        "type": "string"
                    },
                    "payment": {
                        "properties": {
                            "id": {
                                "description": "ID of the payment",
                                "type": "integer"
                            },
                            "value": {
                                "description": "Value of the payment",
                                "type": "number",
                                "format": "float"
                            },
                            "dateTime": {
                                "description": "Date and time when the payment was registered",
                                "type": "string",
                                "format": "date-time"
                            },
                            "paymentMethod": {
                                "description": "Method used for the payment",
                                "type": "string"
                            },
                            "paid": {
                                "description": "Indicates if the payment has been made",
                                "type": "boolean"
                            },
                            "vehicle_id": {
                                "description": "ID of the vehicle associated with the payment",
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "GetAllPaymentsResource": {
                "title": "Get All Payments Resource",
                "description": "Resource returned when retrieving all payments",
                "properties": {
                    "status": {
                        "description": "Indicates if the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "A message describing the result of the operation",
                        "type": "string"
                    },
                    "payments": {
                        "description": "List of payments",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID of the payment",
                                    "type": "integer"
                                },
                                "value": {
                                    "description": "Value of the payment",
                                    "type": "number",
                                    "format": "float"
                                },
                                "dateTime": {
                                    "description": "Date and time when the payment was registered",
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "paymentMethod": {
                                    "description": "Method used for the payment",
                                    "type": "string"
                                },
                                "paid": {
                                    "description": "Indicates if the payment has been made",
                                    "type": "boolean"
                                },
                                "vehicle_id": {
                                    "description": "ID of the vehicle associated with the payment",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "GetPaymentByIdResource": {
                "title": "Get Payment By ID Resource",
                "description": "Resource returned when retrieving a payment by its ID",
                "properties": {
                    "status": {
                        "description": "Indicates if the operation was successful",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "List of errors, if any",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "A message describing the result of the operation",
                        "type": "string"
                    },
                    "payment": {
                        "description": "Details of the payment, which is null if not found",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token to authenticate"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}